+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## v0.2.4 -  2017-05-01
+### Added
+ - Dictionaries endpoint
+ - Comments: create, edit, delete
+
 ## v0.2.3 - 2017-04-29
 ### Fixed:
  - Resume edit